refactor: initial commit, move language detector and time detector to two separate modules (#19)
